feat: rename button, create outline & ghost button
