+++ src/index.js
@@ -1,2 +1,2 @@
 export const version = "0.0.1";
-import { SemaButton } from "./components/SemaButton/sema-button.js";
+import { SemaButton } from "./components/SemaButton/SemaButton.js";
